+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 1. (New) Add initial core logic. (Issue #1)
 1. (New) Add initial CLI. (Issue #4)
+1. (New) Write results of each iteration to a separate file.  (Issue #5)
 1. (New) Add `-s` option to stop on first failure.  (Issue #6)